[WebRTC] Fix logic when parsing H264 packets to make sure the buffer isn't exceeded https://bugs.webkit.org/show_bug.cgi?id=266653 rdar://118861473
Reviewed by NOBODY (OOPS!).

`NextAggregatePacket` can have a free capacity for the buffer that is less than the payload
size of the packet. Added return false to `NextPacket` if this is the case, similar to if
there are no packets to send. `NextFragmentPacket` resets the free capacity so this only
applies if the packets are aggregated

* Source/ThirdParty/libwebrtc/Source/webrtc/modules/rtp_rtcp/source/rtp_format_h264.cc: