Adding preference for controlling font fallback for PUA characters https://bugs.webkit.org/show_bug.cgi?id=266664 rdar://119517735
Reviewed by NOBODY (OOPS!).

WebKit currently complies to CSS fonts specification https://drafts.csswg.org/css-fonts-4/#char-handling-issues
on the fallback behavior for PUA characters. This was implemented at https://bugs.webkit.org/show_bug.cgi?id=263261

The specification requires that the engine doesn't fallback to
system fonts if a character belongs to the Private Use Area (PUA) block.
If the author specifies a font that can render such character, we, of course
keep resolving that glyph accordingly, but if no font specified cover can do that
we won't fallback to a system font.

This behavior breaks the render of certain code points for applications
that have traditionally relied on it. For example, the codepoint
U+F8FF, which is in the PUA block, is mapped to the Apple logo in some fonts
like 'Times' and 'Helvetica' on Apple platforms. Some authors then  might expect
that this works for 'Times New Roman', although the font doesn't support it,
since they are not aware that under the hood WebKit would fallback to a system
font for rendering such a glyph.

For that reason, we are creating a prefererence "DisallowSystemFontFallbackForPrivateUseAreaCharacters"
that allows the application to choose if they want or not to comply to
such CSS Font specification.

When "DisallowSystemFontFallbackForPrivateUseAreaCharacters" is true,
as the name suggests, we won't allow PUA characters to fallback
to system fonts.

* LayoutTests/fast/css/font-fallback-private-use-area-characters-allow-system-fallback-expected.html: Added.
* LayoutTests/fast/css/font-fallback-private-use-area-characters-allow-system-fallback.html: Added.
* LayoutTests/fast/css/font-fallback-private-use-area-characters-disallow-system-fallback-expected-mismatch.html: Added.
* LayoutTests/fast/css/font-fallback-private-use-area-characters-disallow-system-fallback.html: Added.
* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WTF/wtf/cocoa/RuntimeApplicationChecksCocoa.cpp:
(WTF::computeSDKAlignedBehaviors):
* Source/WTF/wtf/cocoa/RuntimeApplicationChecksCocoa.h:
* Source/WebCore/platform/graphics/FontCascadeDescription.h:
(WebCore::FontCascadeDescription::disallowSystemFontFallbackForPrivateUseAreaCharacters const):
(WebCore::FontCascadeDescription::setDisallowSystemFontFallbackForPrivateUseAreaCharacters):
(WebCore::FontCascadeDescription::initialDisallowSystemFontFallbackForPrivateUseAreaCharacters):
(WebCore::FontCascadeDescription::operator== const):
* Source/WebCore/platform/graphics/FontCascadeFonts.cpp:
(WebCore::FontCascadeFonts::glyphDataForVariant):
* Source/WebCore/style/StyleResolveForDocument.cpp:
(WebCore::Style::resolveForDocument):
* Source/WebCore/style/StyleResolver.cpp:
(WebCore::Style::Resolver::defaultStyleForElement):
* Source/WebKit/Shared/WebPreferencesDefaultValues.cpp:
(WebKit::defaultDisallowSystemFontFallbackForPrivateUseAreaCharacters):
* Source/WebKit/Shared/WebPreferencesDefaultValues.h:
* Source/WebKitLegacy/mac/WebView/WebPreferencesDefaultValues.h:
* Source/WebKitLegacy/mac/WebView/WebPreferencesDefaultValues.mm:
(WebKit::defaultDisallowSystemFontFallbackForPrivateUseAreaCharacters):
* Tools/WebKitTestRunner/TestOptions.cpp:
(WTR::TestOptions::defaults):