Minor cleanup in SelectorChecker::matchRecursively() https://bugs.webkit.org/show_bug.cgi?id=266641
Reviewed by Antti Koivisto.

Since we already know we are matching a pseudo element there is no need
for the additional check isCustomPseudoElement() makes.

And then since we are in the condition where
isWebKitCustomPseudoElement() is true, there's no need to check it
again.

* Source/WebCore/css/SelectorChecker.cpp:
(WebCore::SelectorChecker::matchRecursively const):

Canonical link: https://commits.webkit.org/272339@main