Don't error out if we successfully loaded an image
https://bugs.webkit.org/show_bug.cgi?id=266676

Reviewed by NOBODY (OOPS!).

This patch aligns WebKit with Gecko / Firefox and Blink / Chromium.

Merge: https://chromium.googlesource.com/chromium/blink/+/c1fdb4cd2099a27ae10dba405fe2869121c50545

Our current behavior was to show 'fallback' image for http response code being > 400.
While the spec shows that this behavior is OK for the `object` element but not for `img`
element, where the spec is silent and implies that the browser should load and display
image even if the http response was a 404.

* Source/WebCore/html/HTMLImageLoader.cpp:
(HTMLImageLoader::notifyFinished):
* LayoutTests/http/tests/misc/image-error.html: Rebaselined
* LayoutTests/http/tests/misc/image-error-expected.txt: Deleted to make 'ref' test
* LayoutTests/http/tests/misc/image-error-expected.html: New Test Expectation