Fix compact allocators with heap identifiers when MALLOC_HEAP_BREAKDOWN is enabled https://bugs.webkit.org/show_bug.cgi?id=266667 rdar://119891833
Reviewed by Mark Lam.

Adds definitions for DECLARE/DEFINE_COMPACT_ALLOCATOR_WITH_HEAP_IDENTIFIER
when MALLOC_HEAP_BREAKDOWN is enabled, and ensures they are used properly.
Currently, this takes the form of simply invoking the non-compact forms of
the respective macros, which is consistent with our compact allocator paths
elsewhere in that we don't do allocation profiling when we use the system
malloc. Also declares and defines a heap identifier for
WeakPtrImplBaseSingleThread, which was erroneously being used without being
declared.

* Source/WTF/wtf/DebugHeap.h:
* Source/WTF/wtf/FastMalloc.h:
* Source/WTF/wtf/WeakPtr.cpp:
* Source/WTF/wtf/WeakRef.h:

Canonical link: https://commits.webkit.org/272360@main