REGRESSION (268779@main): Style sheets fail to load in the 交通银行 app on iOS https://bugs.webkit.org/show_bug.cgi?id=266716 rdar://118649749
Reviewed by Chris Dumez.

Add a new linked-on-or-after check to guard the (spec-compliant) behavior change in 268779@main. In
particular, the response status code for the stylesheets loaded in this app are 0, due to the
returned `NSURLResponse` in `WKNetworkSessionDelegate` not being a subclass of `NSHTTPURLResponse`.

* Source/WTF/wtf/cocoa/RuntimeApplicationChecksCocoa.h:
* Source/WebCore/loader/cache/CachedCSSStyleSheet.cpp:
(WebCore::CachedCSSStyleSheet::canUseSheet const):

Canonical link: https://commits.webkit.org/272379@main