Make useRandomizingExecutableIslandAllocation do something when using libpas https://bugs.webkit.org/show_bug.cgi?id=266645 rdar://119876990
Reviewed by Alexey Shvayka.

Make useRandomizingExecutableIslandAllocation do something when using libpas.
Before enabling the libpas jit heap, this options randomized allocations of
jit memory, allowing us to stress-test jump islands.

This patch adds in some fragmentation. It is not as comprehensive as the original
version of this option, but it is sufficient to trigger the creation of many
jump islands while running JS2.

* Source/JavaScriptCore/jit/ExecutableAllocator.cpp:
* Source/WTF/wtf/PlatformEnable.h:

Canonical link: https://commits.webkit.org/272434@main