[ New Test ] (264228@main): [ iOS ] imported/w3c/web-platform-tests/html/semantics/disabled-elements/disabled-event-dispatch.tentative.html is a consistent failure https://bugs.webkit.org/show_bug.cgi?id=259132 rdar://112109419
Unreviewed test expectations gardening.

This test failure comprises of both valid and invalid expectations. The
invalid expectation stems from the fact that we never dispatch double
click events on iOS.

The mismatch in valid expectations was fixed by 264098@main, so all this
patch does is clean up the test expectations. Note that this failure
would not have persisted if the expectations gardening in 271716@main was
applied to a `platform/ios/` directory.

* LayoutTests/platform/ios-simulator-16-wk2/imported/w3c/web-platform-tests/html/semantics/disabled-elements/disabled-event-dispatch.tentative-expected.txt: Removed.
* LayoutTests/platform/ios/TestExpectations:
* LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/disabled-elements/disabled-event-dispatch.tentative-expected.txt:

Canonical link: https://commits.webkit.org/272540@main