Introduce run-webkit-tests --site-isolation https://bugs.webkit.org/show_bug.cgi?id=267028
Reviewed by Pascoe.

This makes a convenient way to run tests in a cross-origin iframe with and without
site isolation enabled.  Any output difference is an indication that something needs
to be fixed.  I introduce a new TestExpectations file to allow us to see progress
as we fix things.

* LayoutTests/platform/mac-site-isolation/TestExpectations: Added.
* Tools/Scripts/webkitpy/layout_tests/run_webkit_tests.py:
(parse_args):

Canonical link: https://commits.webkit.org/272619@main